+++ ext/opentelemetry-ext-jinja2/CHANGELOG.md
@@ -1,3 +1,5 @@
 # Changelog
 
 ## Unreleased
+
+- Add jinja2 instrumentation ([#643](https://github.com/open-telemetry/opentelemetry-python/pull/643))